Refactor domain logic into SRP modules
