+++ crates/perfgate/src/domain/mod.rs
@@ -8,8 +8,11 @@
 
 mod blame;
 pub mod budget;
+mod compare;
 pub mod host;
+mod metrics;
 pub mod paired;
+mod report;
 pub mod scaling;
 pub mod significance;
 pub mod stats;
@@ -27,6 +30,14 @@ pub use budget::{
     BudgetError, BudgetResult, aggregate_verdict, calculate_regression, determine_status,
     evaluate_budget, evaluate_budgets, reason_token,
 };
+pub use compare::{
+    Comparison, SignificancePolicy, compare_runs, compare_runs_with_tradeoffs, compare_stats,
+    compare_stats_with_tradeoffs,
+};
+pub use metrics::{compute_stats, metric_value};
+#[cfg(test)]
+pub(crate) use metrics::{metric_cv, metric_to_string, metric_value_from_run};
+pub use report::{Finding, FindingData, Report, derive_report};
 
 pub use significance::{compute_significance, mean_and_variance};
 pub use stats::{
@@ -35,15 +46,16 @@ pub use stats::{
     summarize_f64, summarize_u64,
 };
 
+pub use perfgate_types::error::StatsError;
+
+#[cfg(test)]
 use perfgate_types::{
-    Budget, CHECK_ID_BUDGET, CompareReceipt, Delta, FINDING_CODE_METRIC_FAIL,
-    FINDING_CODE_METRIC_WARN, Metric, MetricStatistic, MetricStatus, RunReceipt, Stats,
+    Budget, CompareReceipt, Delta, Metric, MetricStatistic, MetricStatus, RunReceipt, Stats,
     TradeoffDowngrade, TradeoffRule, VERDICT_REASON_TRADEOFF_MISSING_REQUIRED_METRIC,
     VERDICT_REASON_TRADEOFF_RULE_NOT_SATISFIED, Verdict, VerdictCounts, VerdictStatus,
 };
+#[cfg(test)]
 use std::collections::BTreeMap;
-
-pub use perfgate_types::error::StatsError;
 
 #[derive(Debug, thiserror::Error)]
 pub enum DomainError {
@@ -575,855 +587,6 @@ mod tradeoff_tests {
                 .contains(&VERDICT_REASON_TRADEOFF_MISSING_REQUIRED_METRIC.to_string())
         );
     }
-}
-
-/// Compute perfgate stats from samples.
-///
-/// Warmup samples (`sample.warmup == true`) are excluded.
-///
-/// # Examples
-///
-/// ```
-/// use perfgate::domain::compute_stats;
-/// use perfgate_types::Sample;
-///
-/// let samples = vec![
-///     Sample {
-///         wall_ms: 100, exit_code: 0, warmup: false, timed_out: false,
-///         cpu_ms: None, page_faults: None, ctx_switches: None,
-///         max_rss_kb: None, io_read_bytes: None, io_write_bytes: None,
-///         network_packets: None, energy_uj: None, binary_bytes: None, stdout: None, stderr: None,
-///     },
-///     Sample {
-///         wall_ms: 120, exit_code: 0, warmup: false, timed_out: false,
-///         cpu_ms: None, page_faults: None, ctx_switches: None,
-///         max_rss_kb: None, io_read_bytes: None, io_write_bytes: None,
-///         network_packets: None, energy_uj: None, binary_bytes: None, stdout: None, stderr: None,
-///     },
-/// ];
-///
-/// let stats = compute_stats(&samples, None).unwrap();
-/// assert_eq!(stats.wall_ms.min, 100);
-/// assert_eq!(stats.wall_ms.max, 120);
-/// ```
-#[must_use = "pure computation; call site should use the returned Stats"]
-pub fn compute_stats(
-    samples: &[perfgate_types::Sample],
-    work_units: Option<u64>,
-) -> Result<Stats, DomainError> {
-    let measured: Vec<&perfgate_types::Sample> = samples.iter().filter(|s| !s.warmup).collect();
-    if measured.is_empty() {
-        return Err(DomainError::NoSamples);
-    }
-
-    let wall: Vec<u64> = measured.iter().map(|s| s.wall_ms).collect();
-    let wall_ms = summarize_u64(&wall)?;
-
-    let cpu_vals: Vec<u64> = measured.iter().filter_map(|s| s.cpu_ms).collect();
-    let cpu_ms = if cpu_vals.is_empty() {
-        None
-    } else {
-        Some(summarize_u64(&cpu_vals)?)
-    };
-
-    let page_fault_vals: Vec<u64> = measured.iter().filter_map(|s| s.page_faults).collect();
-    let page_faults = if page_fault_vals.is_empty() {
-        None
-    } else {
-        Some(summarize_u64(&page_fault_vals)?)
-    };
-
-    let ctx_switch_vals: Vec<u64> = measured.iter().filter_map(|s| s.ctx_switches).collect();
-    let ctx_switches = if ctx_switch_vals.is_empty() {
-        None
-    } else {
-        Some(summarize_u64(&ctx_switch_vals)?)
-    };
-
-    let rss_vals: Vec<u64> = measured.iter().filter_map(|s| s.max_rss_kb).collect();
-    let max_rss_kb = if rss_vals.is_empty() {
-        None
-    } else {
-        Some(summarize_u64(&rss_vals)?)
-    };
-
-    let io_read_vals: Vec<u64> = measured.iter().filter_map(|s| s.io_read_bytes).collect();
-    let io_read_bytes = if io_read_vals.is_empty() {
-        None
-    } else {
-        Some(summarize_u64(&io_read_vals)?)
-    };
-
-    let io_write_vals: Vec<u64> = measured.iter().filter_map(|s| s.io_write_bytes).collect();
-    let io_write_bytes = if io_write_vals.is_empty() {
-        None
-    } else {
-        Some(summarize_u64(&io_write_vals)?)
-    };
-
-    let network_vals: Vec<u64> = measured.iter().filter_map(|s| s.network_packets).collect();
-    let network_packets = if network_vals.is_empty() {
-        None
-    } else {
-        Some(summarize_u64(&network_vals)?)
-    };
-
-    let energy_vals: Vec<u64> = measured.iter().filter_map(|s| s.energy_uj).collect();
-    let energy_uj = if energy_vals.is_empty() {
-        None
-    } else {
-        Some(summarize_u64(&energy_vals)?)
-    };
-
-    let binary_vals: Vec<u64> = measured.iter().filter_map(|s| s.binary_bytes).collect();
-    let binary_bytes = if binary_vals.is_empty() {
-        None
-    } else {
-        Some(summarize_u64(&binary_vals)?)
-    };
-
-    let throughput_per_s = match work_units {
-        Some(work) => {
-            let thr: Vec<f64> = measured
-                .iter()
-                .map(|s| {
-                    let secs = (s.wall_ms as f64) / 1000.0;
-                    if secs <= 0.0 {
-                        0.0
-                    } else {
-                        (work as f64) / secs
-                    }
-                })
-                .collect();
-            Some(summarize_f64(&thr)?)
-        }
-        None => None,
-    };
-
-    Ok(Stats {
-        wall_ms,
-        cpu_ms,
-        page_faults,
-        ctx_switches,
-        max_rss_kb,
-        io_read_bytes,
-        io_write_bytes,
-        network_packets,
-        energy_uj,
-        binary_bytes,
-        throughput_per_s,
-    })
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct Comparison {
-    pub deltas: BTreeMap<Metric, Delta>,
-    pub verdict: Verdict,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct SignificancePolicy {
-    pub alpha: f64,
-    pub min_samples: usize,
-    pub require_significance: bool,
-}
-
-impl SignificancePolicy {
-    pub fn new(
-        alpha: f64,
-        min_samples: usize,
-        require_significance: bool,
-    ) -> Result<Self, DomainError> {
-        if !(0.0..=1.0).contains(&alpha) {
-            return Err(DomainError::InvalidAlpha(alpha));
-        }
-        Ok(Self {
-            alpha,
-            min_samples,
-            require_significance,
-        })
-    }
-}
-
-fn aggregate_verdict_from_counts(counts: VerdictCounts, reasons: Vec<String>) -> Verdict {
-    let status = if counts.fail > 0 {
-        VerdictStatus::Fail
-    } else if counts.warn > 0 {
-        VerdictStatus::Warn
-    } else if counts.pass > 0 {
-        VerdictStatus::Pass
-    } else {
-        VerdictStatus::Skip
-    };
-
-    Verdict {
-        status,
-        counts,
-        reasons,
-    }
-}
-
-/// Compare stats under the provided budgets.
-///
-/// Metrics without both baseline+current values are skipped (and therefore do not affect verdict).
-///
-/// # Examples
-///
-/// ```
-/// use perfgate::domain::compare_stats;
-/// use perfgate_types::*;
-/// use std::collections::BTreeMap;
-///
-/// let baseline = Stats {
-///     wall_ms: U64Summary::new(100, 90, 110 ),
-///     cpu_ms: None, page_faults: None, ctx_switches: None,
-///     max_rss_kb: None,
-///     io_read_bytes: None, io_write_bytes: None, network_packets: None,
-///     energy_uj: None,
-///     binary_bytes: None, throughput_per_s: None,
-/// };
-/// let current = Stats {
-///     wall_ms: U64Summary::new(105, 95, 115 ),
-///     cpu_ms: None, page_faults: None, ctx_switches: None,
-///     max_rss_kb: None,
-///     io_read_bytes: None, io_write_bytes: None, network_packets: None,
-///     energy_uj: None,
-///     binary_bytes: None, throughput_per_s: None,
-/// };
-///
-/// let mut budgets = BTreeMap::new();
-/// budgets.insert(Metric::WallMs, Budget {
-///     noise_threshold: None,
-///     noise_policy: perfgate_types::NoisePolicy::Ignore,
-///     threshold: 0.20, warn_threshold: 0.10, direction: Direction::Lower,
-/// });
-///
-/// let cmp = compare_stats(&baseline, &current, &budgets).unwrap();
-/// assert_eq!(cmp.verdict.status, VerdictStatus::Pass);
-/// ```
-#[must_use = "pure computation; call site should use the returned Comparison"]
-pub fn compare_stats(
-    baseline: &Stats,
-    current: &Stats,
-    budgets: &BTreeMap<Metric, Budget>,
-) -> Result<Comparison, DomainError> {
-    compare_stats_with_tradeoffs(baseline, current, budgets, &[])
-}
-
-/// Compare stats under the provided budgets and optional tradeoff rules.
-#[must_use = "pure computation; call site should use the returned Comparison"]
-pub fn compare_stats_with_tradeoffs(
-    baseline: &Stats,
-    current: &Stats,
-    budgets: &BTreeMap<Metric, Budget>,
-    tradeoffs: &[TradeoffRule],
-) -> Result<Comparison, DomainError> {
-    let mut deltas: BTreeMap<Metric, Delta> = BTreeMap::new();
-    let mut reasons: Vec<String> = Vec::new();
-
-    let mut counts = VerdictCounts {
-        pass: 0,
-        warn: 0,
-        fail: 0,
-        skip: 0,
-    };
-
-    for (metric, budget) in budgets {
-        let b = metric_value(baseline, *metric);
-        let c = metric_value(current, *metric);
-        let current_cv = metric_cv(current, *metric);
-
-        let (Some(bv), Some(cv)) = (b, c) else {
-            continue;
-        };
-
-        if bv <= 0.0 {
-            deltas.insert(
-                *metric,
-                Delta {
-                    baseline: bv,
-                    current: cv,
-                    ratio: 1.0,
-                    pct: 0.0,
-                    regression: 0.0,
-                    status: MetricStatus::Skip,
-                    significance: None,
-                    cv: current_cv,
-                    noise_threshold: budget.noise_threshold,
-                    statistic: MetricStatistic::Median,
-                },
-            );
-            counts.skip += 1;
-            continue;
-        }
-
-        let result = evaluate_budget(bv, cv, budget, current_cv)
-            .expect("evaluate_budget is infallible for bv > 0");
-
-        match result.status {
-            MetricStatus::Pass => counts.pass += 1,
-            MetricStatus::Warn => {
-                counts.warn += 1;
-                reasons.push(reason_token(*metric, MetricStatus::Warn));
-            }
-            MetricStatus::Fail => {
-                counts.fail += 1;
-                reasons.push(reason_token(*metric, MetricStatus::Fail));
-            }
-            MetricStatus::Skip => {
-                counts.skip += 1;
-                reasons.push(reason_token(*metric, MetricStatus::Skip));
-            }
-        }
-
-        deltas.insert(
-            *metric,
-            Delta {
-                baseline: result.baseline,
-                current: result.current,
-                ratio: result.ratio,
-                pct: result.pct,
-                regression: result.regression,
-                cv: result.cv,
-                noise_threshold: result.noise_threshold,
-                statistic: MetricStatistic::Median,
-                significance: None,
-                status: result.status,
-            },
-        );
-    }
-
-    apply_tradeoffs(&mut deltas, &mut counts, &mut reasons, tradeoffs);
-    let verdict = aggregate_verdict_from_counts(counts, reasons);
-
-    Ok(Comparison { deltas, verdict })
-}
-
-/// Compare full run receipts under the provided budgets.
-///
-/// This variant supports:
-/// - Per-metric statistic selection (`median` or `p95`)
-/// - Optional significance analysis with Welch's t-test
-pub fn compare_runs(
-    baseline: &RunReceipt,
-    current: &RunReceipt,
-    budgets: &BTreeMap<Metric, Budget>,
-    metric_statistics: &BTreeMap<Metric, MetricStatistic>,
-    significance_policy: Option<SignificancePolicy>,
-) -> Result<Comparison, DomainError> {
-    compare_runs_with_tradeoffs(
-        baseline,
-        current,
-        budgets,
-        metric_statistics,
-        significance_policy,
-        &[],
-    )
-}
-
-/// Compare full run receipts under budgets with optional tradeoff rules.
-pub fn compare_runs_with_tradeoffs(
-    baseline: &RunReceipt,
-    current: &RunReceipt,
-    budgets: &BTreeMap<Metric, Budget>,
-    metric_statistics: &BTreeMap<Metric, MetricStatistic>,
-    significance_policy: Option<SignificancePolicy>,
-    tradeoffs: &[TradeoffRule],
-) -> Result<Comparison, DomainError> {
-    let mut deltas: BTreeMap<Metric, Delta> = BTreeMap::new();
-    let mut reasons: Vec<String> = Vec::new();
-
-    let mut counts = VerdictCounts {
-        pass: 0,
-        warn: 0,
-        fail: 0,
-        skip: 0,
-    };
-
-    for (metric, budget) in budgets {
-        let statistic = metric_statistics
-            .get(metric)
-            .copied()
-            .unwrap_or(MetricStatistic::Median);
-
-        let b = metric_value_from_run(baseline, *metric, statistic);
-        let c = metric_value_from_run(current, *metric, statistic);
-        let current_cv = metric_cv(&current.stats, *metric);
-
-        let (Some(bv), Some(cv)) = (b, c) else {
-            continue;
-        };
-
-        if bv <= 0.0 {
-            deltas.insert(
-                *metric,
-                Delta {
-                    baseline: bv,
-                    current: cv,
-                    ratio: 1.0,
-                    pct: 0.0,
-                    regression: 0.0,
-                    status: MetricStatus::Skip,
-                    significance: None,
-                    cv: current_cv,
-                    noise_threshold: budget.noise_threshold,
-                    statistic,
-                },
-            );
-            counts.skip += 1;
-            continue;
-        }
-
-        let result = evaluate_budget(bv, cv, budget, current_cv)
-            .expect("evaluate_budget is infallible for bv > 0");
-
-        let mut status = result.status;
-
-        let significance = significance_policy.and_then(|policy| {
-            let baseline_series = metric_series_from_run(baseline, *metric);
-            let current_series = metric_series_from_run(current, *metric);
-            compute_significance(
-                &baseline_series,
-                &current_series,
-                policy.alpha,
-                policy.min_samples,
-            )
-        });
-
-        if let Some(policy) = significance_policy
-            && policy.require_significance
-            && matches!(status, MetricStatus::Warn | MetricStatus::Fail)
-        {
-            let is_significant = significance
-                .as_ref()
-                .map(|sig| sig.significant)
-                .unwrap_or(false);
-            if !is_significant {
-                status = MetricStatus::Pass;
-            }
-        }
-
-        match status {
-            MetricStatus::Pass => counts.pass += 1,
-            MetricStatus::Warn => {
-                counts.warn += 1;
-                reasons.push(reason_token(*metric, MetricStatus::Warn));
-            }
-            MetricStatus::Fail => {
-                counts.fail += 1;
-                reasons.push(reason_token(*metric, MetricStatus::Fail));
-            }
-            MetricStatus::Skip => {
-                counts.skip += 1;
-                reasons.push(reason_token(*metric, MetricStatus::Skip));
-            }
-        }
-
-        deltas.insert(
-            *metric,
-            Delta {
-                baseline: result.baseline,
-                current: result.current,
-                ratio: result.ratio,
-                pct: result.pct,
-                regression: result.regression,
-                cv: result.cv,
-                noise_threshold: result.noise_threshold,
-                statistic,
-                significance,
-                status,
-            },
-        );
-    }
-
-    apply_tradeoffs(&mut deltas, &mut counts, &mut reasons, tradeoffs);
-    let verdict = aggregate_verdict_from_counts(counts, reasons);
-
-    Ok(Comparison { deltas, verdict })
-}
-
-fn push_unique_reason(reasons: &mut Vec<String>, token: String) {
-    if !reasons.contains(&token) {
-        reasons.push(token);
-    }
-}
-
-fn remove_reason(reasons: &mut Vec<String>, token: &str) {
-    reasons.retain(|reason| reason != token);
-}
-
-fn improvement_ratio(delta: &Delta, metric: Metric) -> Option<f64> {
-    match metric.default_direction() {
-        perfgate_types::Direction::Higher => Some(delta.ratio),
-        perfgate_types::Direction::Lower => Some(if delta.current <= 0.0 {
-            f64::INFINITY
-        } else {
-            delta.baseline / delta.current
-        }),
-    }
-}
-
-fn apply_tradeoffs(
-    deltas: &mut BTreeMap<Metric, Delta>,
-    counts: &mut VerdictCounts,
-    reasons: &mut Vec<String>,
-    tradeoffs: &[TradeoffRule],
-) {
-    let failed_metrics: Vec<Metric> = deltas
-        .iter()
-        .filter_map(|(metric, delta)| (delta.status == MetricStatus::Fail).then_some(*metric))
-        .collect();
-
-    for failed_metric in failed_metrics {
-        let mut applied = false;
-        let mut saw_missing_required_metric = false;
-        let mut saw_unsatisfied_rule = false;
-
-        for rule in tradeoffs
-            .iter()
-            .filter(|rule| rule.if_failed == failed_metric)
-        {
-            let mut missing_required_metric = false;
-            let mut satisfied = !rule.require.is_empty();
-
-            for requirement in &rule.require {
-                if requirement.probe.is_some() {
-                    missing_required_metric = true;
-                    satisfied = false;
-                    break;
-                }
-
-                let Some(required_delta) = deltas.get(&requirement.metric) else {
-                    missing_required_metric = true;
-                    satisfied = false;
-                    break;
-                };
-
-                let Some(ratio) = improvement_ratio(required_delta, requirement.metric) else {
-                    satisfied = false;
-                    break;
-                };
-
-                if ratio < requirement.min_improvement_ratio {
-                    satisfied = false;
-                }
-            }
-
-            if !rule.allow.is_empty() {
-                missing_required_metric = true;
-                satisfied = false;
-            }
-
-            if !satisfied {
-                saw_missing_required_metric |= missing_required_metric;
-                saw_unsatisfied_rule = true;
-                continue;
-            }
-
-            if let Some(delta) = deltas.get_mut(&failed_metric) {
-                let new_status = match rule.downgrade_to {
-                    TradeoffDowngrade::Warn => MetricStatus::Warn,
-                    TradeoffDowngrade::Pass => MetricStatus::Pass,
-                };
-                if delta.status == MetricStatus::Fail && new_status != MetricStatus::Fail {
-                    counts.fail = counts.fail.saturating_sub(1);
-                    match new_status {
-                        MetricStatus::Warn => counts.warn += 1,
-                        MetricStatus::Pass => counts.pass += 1,
-                        MetricStatus::Fail | MetricStatus::Skip => {}
-                    }
-                    delta.status = new_status;
-                    remove_reason(reasons, &reason_token(failed_metric, MetricStatus::Fail));
-                    if new_status == MetricStatus::Warn {
-                        push_unique_reason(
-                            reasons,
-                            reason_token(failed_metric, MetricStatus::Warn),
-                        );
-                    }
-                }
-            }
-
-            push_unique_reason(reasons, format!("tradeoff_{}_applied", rule.name));
-            applied = true;
-            break;
-        }
-
-        if !applied {
-            if saw_missing_required_metric {
-                push_unique_reason(
-                    reasons,
-                    VERDICT_REASON_TRADEOFF_MISSING_REQUIRED_METRIC.to_string(),
-                );
-            }
-            if saw_unsatisfied_rule {
-                push_unique_reason(
-                    reasons,
-                    VERDICT_REASON_TRADEOFF_RULE_NOT_SATISFIED.to_string(),
-                );
-            }
-        }
-    }
-}
-
-// ============================================================================
-// Report Derivation
-// ============================================================================
-
-/// Data for a single finding in a report.
-#[derive(Debug, Clone, PartialEq)]
-pub struct FindingData {
-    /// The metric name (e.g., "wall_ms", "max_rss_kb", "throughput_per_s").
-    pub metric_name: String,
-    /// The benchmark name.
-    pub bench_name: String,
-    /// Baseline value for the metric.
-    pub baseline: f64,
-    /// Current value for the metric.
-    pub current: f64,
-    /// Regression percentage (e.g., 0.15 means 15% regression).
-    pub regression_pct: f64,
-    /// The threshold that was exceeded (for fail) or approached (for warn).
-    pub threshold: f64,
-}
-
-/// A single finding in a report.
-#[derive(Debug, Clone, PartialEq)]
-pub struct Finding {
-    /// Finding code: "metric_warn" or "metric_fail".
-    pub code: String,
-    /// Check identifier: always "perf.budget".
-    pub check_id: String,
-    /// Finding data containing metric details.
-    pub data: FindingData,
-}
-
-/// Report derived from a CompareReceipt.
-#[derive(Debug, Clone, PartialEq)]
-pub struct Report {
-    /// The overall verdict status, matching the compare verdict.
-    pub verdict: VerdictStatus,
-    /// Findings for metrics that have Warn or Fail status.
-    /// Ordered deterministically by metric name, then bench name.
-    pub findings: Vec<Finding>,
-}
-
-/// Derives a report from a CompareReceipt.
-///
-/// Creates findings for each metric delta with status Warn or Fail.
-/// Findings are ordered deterministically by metric name (then bench name if
-/// multiple benches were compared, though currently CompareReceipt is per-bench).
-///
-/// # Invariants
-///
-/// - Number of findings equals count of warn + fail status deltas
-/// - Report verdict matches compare verdict
-/// - Findings are ordered deterministically (by metric name)
-///
-/// # Examples
-///
-/// ```
-/// use perfgate::domain::derive_report;
-/// use perfgate_types::*;
-/// use std::collections::BTreeMap;
-///
-/// let receipt = CompareReceipt {
-///     schema: COMPARE_SCHEMA_V1.to_string(),
-///     tool: ToolInfo { name: "perfgate".into(), version: "0.1.0".into() },
-///     bench: BenchMeta {
-///         name: "my-bench".into(), cwd: None,
-///         command: vec!["echo".into()], repeat: 5, warmup: 0,
-///         work_units: None, timeout_ms: None,
-///     },
-///     baseline_ref: CompareRef { path: None, run_id: None },
-///     current_ref: CompareRef { path: None, run_id: None },
-///     budgets: BTreeMap::new(),
-///     deltas: BTreeMap::new(),
-///     verdict: Verdict {
-///         status: VerdictStatus::Pass,
-///         counts: VerdictCounts { pass: 0, warn: 0, fail: 0, skip: 0 },
-///         reasons: vec![],
-///     },
-/// };
-///
-/// let report = derive_report(&receipt);
-/// assert_eq!(report.verdict, VerdictStatus::Pass);
-/// assert!(report.findings.is_empty());
-/// ```
-#[must_use = "pure computation; call site should use the returned Report"]
-pub fn derive_report(receipt: &CompareReceipt) -> Report {
-    let mut findings = Vec::new();
-
-    // Iterate over deltas in deterministic order (BTreeMap is sorted by key)
-    for (metric, delta) in &receipt.deltas {
-        match delta.status {
-            MetricStatus::Pass | MetricStatus::Skip => continue,
-            MetricStatus::Warn | MetricStatus::Fail => {
-                let code = match delta.status {
-                    MetricStatus::Warn => FINDING_CODE_METRIC_WARN.to_string(),
-                    MetricStatus::Fail => FINDING_CODE_METRIC_FAIL.to_string(),
-                    _ => unreachable!(),
-                };
-
-                // Get the threshold from budgets if available
-                let threshold = receipt
-                    .budgets
-                    .get(metric)
-                    .map(|b| b.threshold)
-                    .unwrap_or(0.0);
-
-                findings.push(Finding {
-                    code,
-                    check_id: CHECK_ID_BUDGET.to_string(),
-                    data: FindingData {
-                        metric_name: metric_to_string(*metric),
-                        bench_name: receipt.bench.name.clone(),
-                        baseline: delta.baseline,
-                        current: delta.current,
-                        regression_pct: delta.regression,
-                        threshold,
-                    },
-                });
-            }
-        }
-    }
-
-    // Findings are already sorted by metric name since we iterate over BTreeMap
-    // For multi-bench scenarios (future), we would also sort by bench_name
-    // Currently sorting is: metric name (from BTreeMap order)
-
-    Report {
-        verdict: receipt.verdict.status,
-        findings,
-    }
-}
-
-fn metric_cv(stats: &Stats, metric: Metric) -> Option<f64> {
-    match metric {
-        Metric::BinaryBytes => stats.binary_bytes.as_ref().and_then(|s| s.cv()),
-        Metric::CpuMs => stats.cpu_ms.as_ref().and_then(|s| s.cv()),
-        Metric::CtxSwitches => stats.ctx_switches.as_ref().and_then(|s| s.cv()),
-        Metric::EnergyUj => stats.energy_uj.as_ref().and_then(|s| s.cv()),
-        Metric::IoReadBytes => stats.io_read_bytes.as_ref().and_then(|s| s.cv()),
-        Metric::IoWriteBytes => stats.io_write_bytes.as_ref().and_then(|s| s.cv()),
-        Metric::MaxRssKb => stats.max_rss_kb.as_ref().and_then(|s| s.cv()),
-        Metric::NetworkPackets => stats.network_packets.as_ref().and_then(|s| s.cv()),
-        Metric::PageFaults => stats.page_faults.as_ref().and_then(|s| s.cv()),
-        Metric::ThroughputPerS => stats.throughput_per_s.as_ref().and_then(|s| s.cv()),
-        Metric::WallMs => stats.wall_ms.cv(),
-    }
-}
-
-/// Converts a Metric enum to its string representation.
-fn metric_to_string(metric: Metric) -> String {
-    metric.as_str().to_string()
-}
-
-#[must_use = "pure computation; call site should use the returned value"]
-pub fn metric_value(stats: &Stats, metric: Metric) -> Option<f64> {
-    match metric {
-        Metric::BinaryBytes => stats.binary_bytes.as_ref().map(|s| s.median as f64),
-        Metric::CpuMs => stats.cpu_ms.as_ref().map(|s| s.median as f64),
-        Metric::CtxSwitches => stats.ctx_switches.as_ref().map(|s| s.median as f64),
-        Metric::EnergyUj => stats.energy_uj.as_ref().map(|s| s.median as f64),
-        Metric::IoReadBytes => stats.io_read_bytes.as_ref().map(|s| s.median as f64),
-        Metric::IoWriteBytes => stats.io_write_bytes.as_ref().map(|s| s.median as f64),
-        Metric::MaxRssKb => stats.max_rss_kb.as_ref().map(|s| s.median as f64),
-        Metric::NetworkPackets => stats.network_packets.as_ref().map(|s| s.median as f64),
-        Metric::PageFaults => stats.page_faults.as_ref().map(|s| s.median as f64),
-        Metric::ThroughputPerS => stats.throughput_per_s.as_ref().map(|s| s.median),
-        Metric::WallMs => Some(stats.wall_ms.median as f64),
-    }
-}
-
-fn metric_value_from_run(
-    run: &RunReceipt,
-    metric: Metric,
-    statistic: MetricStatistic,
-) -> Option<f64> {
-    match statistic {
-        MetricStatistic::Median => metric_value(&run.stats, metric),
-        MetricStatistic::P95 => {
-            let values = metric_series_from_run(run, metric);
-            if values.is_empty() {
-                metric_value(&run.stats, metric)
-            } else {
-                percentile(values, 0.95)
-            }
-        }
-    }
-}
-
-fn metric_series_from_run(run: &RunReceipt, metric: Metric) -> Vec<f64> {
-    let measured = run.samples.iter().filter(|s| !s.warmup);
-
-    match metric {
-        Metric::BinaryBytes => measured
-            .filter_map(|s| s.binary_bytes.map(|v| v as f64))
-            .collect(),
-        Metric::CpuMs => measured
-            .filter_map(|s| s.cpu_ms.map(|v| v as f64))
-            .collect(),
-        Metric::CtxSwitches => measured
-            .filter_map(|s| s.ctx_switches.map(|v| v as f64))
-            .collect(),
-        Metric::EnergyUj => measured
-            .filter_map(|s| s.energy_uj.map(|v| v as f64))
-            .collect(),
-        Metric::IoReadBytes => measured
-            .filter_map(|s| s.io_read_bytes.map(|v| v as f64))
-            .collect(),
-        Metric::IoWriteBytes => measured
-            .filter_map(|s| s.io_write_bytes.map(|v| v as f64))
-            .collect(),
-        Metric::MaxRssKb => measured
-            .filter_map(|s| s.max_rss_kb.map(|v| v as f64))
-            .collect(),
-        Metric::NetworkPackets => measured
-            .filter_map(|s| s.network_packets.map(|v| v as f64))
-            .collect(),
-        Metric::PageFaults => measured
-            .filter_map(|s| s.page_faults.map(|v| v as f64))
-            .collect(),
-        Metric::ThroughputPerS => {
-            let Some(work) = run.bench.work_units else {
-                return Vec::new();
-            };
-            measured
-                .map(|s| {
-                    let secs = (s.wall_ms as f64) / 1000.0;
-                    if secs <= 0.0 {
-                        0.0
-                    } else {
-                        (work as f64) / secs
-                    }
-                })
-                .collect()
-        }
-        Metric::WallMs => measured.map(|s| s.wall_ms as f64).collect(),
-    }
-}
-
-fn percentile(mut values: Vec<f64>, q: f64) -> Option<f64> {
-    if values.is_empty() {
-        return None;
-    }
-
-    values.sort_by(|a, b| a.partial_cmp(b).unwrap_or(std::cmp::Ordering::Equal));
-
-    if values.len() == 1 {
-        return Some(values[0]);
-    }
-
-    let rank = q.clamp(0.0, 1.0) * (values.len() as f64 - 1.0);
-    let lower = rank.floor() as usize;
-    let upper = rank.ceil() as usize;
-
-    if lower == upper {
-        return Some(values[lower]);
-    }
-
-    let weight = rank - lower as f64;
-    Some(values[lower] + (values[upper] - values[lower]) * weight)
 }
 
 #[cfg(test)]
